% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keanu-package.R
\name{get_distribution_params}
\alias{get_distribution_params}
\title{Given coefficients (in a data.frame) and a list of model-mats, get predicted parameters for the
distribution}
\usage{
get_distribution_params(par_tbl, model_mats, inv_link_funs = NULL)
}
\arguments{
\item{par_tbl}{A data.frame with parameter, term, beta}

\item{model_mats}{A list of model-matrices}

\item{inv_link_funs}{A named list of inverse-link functions (names must match parameters)}
}
\value{
A data.frame where each row corresponds to the rows of the model-mats, and each column is
  a distribution-parameter.
}
\description{
Given coefficients (in a data.frame) and a list of model-mats, get predicted parameters for the
distribution
}
