% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keanu-package.R
\name{initialize_df_pars}
\alias{initialize_df_pars}
\title{Initialize a dataframe with parameter, term, beta}
\usage{
initialize_df_pars(list_of_model_mats)
}
\arguments{
\item{list_of_model_mats}{A list of model-matrices}
}
\value{
A data.frame with parameter, term, beta
}
\description{
Initialize a dataframe with parameter, term, beta
}
