% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keanu-package.R
\name{initialize_par_tbl}
\alias{initialize_par_tbl}
\title{Initialize a dataframe with parameter, term, beta}
\usage{
initialize_par_tbl(model_mats)
}
\arguments{
\item{model_mats}{A list of model-matrices}
}
\value{
A data.frame with parameter, term, beta
}
\description{
Initialize a dataframe with parameter, term, beta
}
