% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keanu-package.R
\name{inside_parens}
\alias{inside_parens}
\title{Return whatever is inside the parantheses of a string.}
\usage{
inside_parens(x)
}
\arguments{
\item{x}{A character-vector.}
}
\value{
Whatever is inside the parentheses. If no parantheses, return original string.
}
\description{
Return whatever is inside the parantheses of a string.
}
