% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keanu-package.R
\name{keanu_model}
\alias{keanu_model}
\title{Fit a keanu model}
\usage{
keanu_model(forms, data, error_fun, penalty_fun = function(x)
  return(numeric(0)), inv_link_funs = NULL, standardize_x = FALSE,
  na.action = na.warn, contrasts = NULL, inits = NULL,
  optim_args = list(method = "BFGS", control = list(trace =
  as.integer(interactive()), maxit = 250)))
}
\arguments{
\item{forms}{A list of formulas, with names for each parameter/aspect of the target-variable
(e.g., 'mean' and 'sd' for a normally distributed variable where both of these parameters can
depend on covariates). The first formula should have a left-hand side, which specifies the
target-variable, as is typical for R modelling functions. The remaining formulas do not need
left-hand-sides.}

\item{data}{A data.frame}

\item{error_fun}{A function which takes two arguments: (1) a data.frame where the columns are
parameters of the distribution being predicted, and rows correspond to observations, and (2)
the target variable being predicted. It returns something to be minimized: e.g., the error, the
negative-log-likelihood.}

\item{penalty_fun}{A function which takes a data.frame with parameter, term, beta, and returns a
penalty. This is added to the result of \code{error_fun}, so that they are minimized jointly.}

\item{inv_link_funs}{A named list of inverse-link functions: i.e., functions that transform
real-valued predictions into (potentially constrained) parameter-estimates. The list must be
named for each parameter of the distribution. Defaults to the identity function for all.}

\item{standardize_x}{Either a logical specifying whether to center/scale numeric predictors in
the model.frame, or a list with names 'center' and 'scale', each of which in turn are named
numeric vectors specifying centering/scaling. Note that terms that result in matrices (e.g.,
`stats::poly`) will not be scaled/centered, nor will factors.}

\item{na.action}{A function specifying what to do with NAs. Defaults to \code{na.warn}.}

\item{contrasts}{Passed to \code{model.matrix}.}

\item{inits}{Not yet supported.}

\item{optim_args}{A named list that will be passed as arguments to \code{stats::optim}.}
}
\value{
An object of class 'keanu_model', with predict method
}
\description{
Fit a keanu model
}
