% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeller_from_likelihood.R
\name{modeller_from_loglik}
\alias{modeller_from_loglik}
\title{Create your own modelling function, given a log-likelihood function.}
\usage{
modeller_from_loglik(loglik_fun, penalty_fun = function(...) return(0),
  non_parameter_args = character())
}
\arguments{
\item{loglik_fun}{A function whose first argument is the target variable, and whose subsequent
arguments are parameters of the distribution which gives the log-likelihood. (You can indicate
arguments that are *not* parameters with \code{non_parameter_args}.)}

\item{penalty_fun}{Optional. A function which, when given a dataframe with columns \code{c('parameter',
'term', 'beta')}, yields a penalty term (a positive number to be added to the function to be
minimized by \code{stats::optim}).}

\item{non_parameter_args}{A character-vector indicating arguments to \code{loglik_fun} that are
not parameters. (This is in addition to the first argument, which is always assumed to be the
target variable, not a parameter.)}
}
\value{
A modelling function. See \code{modeller_template}.
}
\description{
This function takes a log-likelihood function, and returns a modelling function, much like
\code{stats:lm} or \code{stats:glm}. This modelling function will find coefficients that maximize
the likelihood.
}
