% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeller_from_likelihood.R
\name{modeller_template}
\alias{modeller_template}
\title{A "template" for a modelling function}
\usage{
modeller_template(formulas, data, standardize_x = FALSE,
  na.action = na.warn, loglik_fun_args = list(),
  penalty_fun_args = list(), optim_args = list(method = "BFGS", control =
  list(trace = as.integer(interactive()), maxit = 250)), contrasts = NULL)
}
\arguments{
\item{formulas}{A list of formulas. The LHS should contain the dependent variable wrapped around a parameter-- e.g., \code{mu(y) ~ x}.}

\item{data}{A data.frame.}

\item{standardize_x}{Should data be standardized before passing to optimizer? Can help with convergence. Affects interpretation of coefficients.}

\item{na.action}{Action when records in data are NA}

\item{loglik_fun_args}{Further arguments to pass to the loglik function (from \code{modeller_from_loglik}).}

\item{penalty_fun_args}{Further arguments to pass to the penalty function (from \code{modeller_from_loglik}).}

\item{optim_args}{A list of arguments to pass to \code{stats:optim}.}
}
\value{
An object of class 'keanu_model', with print and predict methods.
}
\description{
This function will not work by itself. Instead, \code{modeller_from_loglik} provides a modified version of it.
}
