% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keanu-package.R
\name{na.warn}
\alias{na.warn}
\title{Handle missing values with a warning}
\usage{
na.warn(object)
}
\arguments{
\item{object}{}
}
\description{
Modified version from \code{modelr}, where \code{warning(immediate.=TRUE)}.
}
