% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keanu-package.R
\name{parse_char}
\alias{parse_char}
\title{Parse a character into an expression}
\usage{
parse_char(x)
}
\arguments{
\item{x}{A character}
}
\value{
An expression
}
\description{
Parse a character into an expression
}
