% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeller_from_likelihood.R
\name{parse_link}
\alias{parse_link}
\title{Parse links: a helper function for `modeller_template`}
\usage{
parse_link(the_link)
}
\arguments{
\item{the_link}{A character naming a link or a named list of functions with elements
'inv_link' and 'link'.}
}
\value{
A named list of functions with elements 'inv_link' and 'link'.
}
\description{
Parse links: a helper function for `modeller_template`
}
