% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keanu-package.R
\name{plot_coefficients}
\alias{plot_coefficients}
\title{Plot coefficients from a model}
\usage{
plot_coefficients(x, terms, se_multi, ...)
}
\arguments{
\item{x}{A model object}

\item{terms}{Terms to include in plot, defaulting to everything but the intercept. This can be a
character-vector, or, more conveniently, a call to \code{dplyr::vars}, which provides a
flexible semantics for including/excluding arbitrary terms.}

\item{se_multi}{Standard-errors will be mulitplied by this to obtain confidence-intervals.}

\item{...}{Arguments to be passed to subsequent methods. Additional arguments will be passed to
`ggplot2::geom_pointrange`}
}
\value{
A ggplot2 object.
}
\description{
Plot coefficients from a model
}
