% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keanu-package.R
\name{remove_lhs}
\alias{remove_lhs}
\title{Remove the LHS of a formula}
\usage{
remove_lhs(x)
}
\arguments{
\item{x}{A formula}
}
\value{
A formula, without its left-hand-side
}
\description{
Remove the LHS of a formula
}
