% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keanu-package.R
\name{roll_df_pars}
\alias{roll_df_pars}
\title{Roll up a flat vector into a data.frame with parameter, term, beta}
\usage{
roll_df_pars(unrolled_par)
}
\arguments{
\item{unrolled_par}{A named numeric vector, named with parameter__sep__term}
}
\value{
A data.frame with parameter, term, beta
}
\description{
Roll up a flat vector into a data.frame with parameter, term, beta
}
