% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keanu-package.R
\name{univariate_predictions}
\alias{univariate_predictions}
\title{Produce univariate predictions}
\usage{
univariate_predictions(object, ...)
}
\arguments{
\item{object}{A model object}
}
\value{
A list of data.frames, one for each column in the model.frame, each consisting of that
  column and the corresponding predictions.
}
\description{
Produce univariate predictions
}
