% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keanu-package.R
\name{unroll_df_pars}
\alias{unroll_df_pars}
\title{Take a data.frame with parameter, term, beta, unroll into a flat vector}
\usage{
unroll_df_pars(df_pars)
}
\arguments{
\item{df_pars}{A data.frame with parameter, term, beta}
}
\value{
A named numeric vector, named with parameter__sep__term
}
\description{
Take a data.frame with parameter, term, beta, unroll into a flat vector
}
