% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keanu-package.R
\name{unroll_par_tbl}
\alias{unroll_par_tbl}
\title{Take a data.frame with parameter, term, beta, unroll into a flat vector}
\usage{
unroll_par_tbl(par_tbl)
}
\arguments{
\item{par_tbl}{A data.frame with parameter, term, beta}
}
\value{
A named numeric vector, named with parameter__sep__term
}
\description{
Take a data.frame with parameter, term, beta, unroll into a flat vector
}
