% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeller_from_likelihood.R
\name{optim_from_mc}
\alias{optim_from_mc}
\title{Given the output of `prepare_model_components`, find coefficient-estimates using `stats::optim`.}
\usage{
optim_from_mc(mc, inv_links, optim_args, loglik_fun, loglik_fun_args,
  penalty_fun, penalty_fun_args)
}
\arguments{
\item{mc}{The output of `prepare_model_components`}

\item{inv_links}{A list of inverse-link-functions, one corresponding to each `mc$model_mats`}

\item{optim_args}{A named list of arguments to be passed to `stats::optim`.}

\item{loglik_fun}{A log-liklihood function to optimize}

\item{loglik_fun_args}{Additional arguments (aside from target and parameters) to pass to the
`loglik_fun`.}

\item{penalty_fun}{A penalty function that takes a table with parameter, term, beta, and returns
a numeric penalty that will be factored into the optimization.}

\item{penalty_fun_args}{Additional arguments (aside from the table) to pass to the `penalty_fun`.}
}
\value{
A list with 'optim' (the results of the call to stats::optim), and 'cov' (the approximate
  covariance matrix obtaining from the hessian).
}
\description{
Given the output of `prepare_model_components`, find coefficient-estimates using `stats::optim`.
}
